doc-requirements.txt
# Requirements for building docs
# Check these dependencies against doc/conf.py
-r dev-requirements.txt
sphinx>=1.4
numpydoc
matplotlib
texext
nb2plots
rpy2
seaborn
statsmodels
tensorflow
keras
nbsphinx
